added migraion changed name of table history

+++ server/UserService/UserService.Api/ApiMappingProfile.cs
@@ -15,8 +15,8 @@ namespace UserService.Api
             CreateMap<EmailVerificationDTO, EmailVerificationModel>();
 
             CreateMap<SucceededOperationModel, SucceededOperationDTO>()
-                .ForMember(dest => dest.Balance, opt => opt.MapFrom(m => m.Balance  / 100))
-                .ForMember(dest => dest.TransactionAmount, opt => opt.MapFrom(m => m.TransactionAmount / 100));
+                .ForMember(dest => dest.Balance, opt => opt.MapFrom(m => (float)m.Balance  / 100  ))
+                .ForMember(dest => dest.TransactionAmount, opt => opt.MapFrom(m => (float)m.TransactionAmount / 100));
 
             CreateMap<PaginationParamsDTO, PaginationParamsModel>();
